added trajectory planner

+++ src/include/math_tools.h
@@ -3,6 +3,6 @@
 #include <iostream>
 
 
-Eigen::VectorXd fifthOrderPolynomialTrajectory(double tf, double start_q, double end_q);
-Eigen::VectorXd fifthOrderPolynomialTrajectory(double tf, double start_q, double start_v, double start_a, double end_q, double end_v, double end_a);
+Eigen::VectorXd fifthOrderPolynomialTrajectory(float tf, double start_q, double end_q);
+Eigen::VectorXd fifthOrderPolynomialTrajectory(float tf, double start_q, double start_v, double start_a, double end_q, double end_v, double end_a);
 Eigen::Matrix3d euler_to_rotation_matrix(Eigen::Vector3d rpy);
